requirements.txt
alembic==1.4.0
aniso8601==8.0.0
arrow==0.15.5
attrs==19.3.0
bcrypt==3.1.7
blinker==1.4
Bootstrap-Flask==1.2.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
dominate==2.4.0
elasticsearch==7.10.0
elasticsearch-dsl==7.1.0
Faker==4.0.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
idna==2.8
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
libgravatar==0.2.3
Mako==1.1.1
MarkupSafe==1.1.1
phonenumbers==8.11.3
pluggy==0.13.1
pycparser==2.19
pycryptodomex==3.9.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyrsistent==0.15.7
python-alipay-sdk==2.0.1
python-dateutil==2.8.1
python-dotenv==0.10.5
python-editor==1.0.4
pytz==2019.3
redis==3.4.1
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.13
text-unidecode==1.3
urllib3==1.25.8
visitor==0.1.3
Werkzeug==0.16.1
WTForms==2.2.1
zipp==2.1.0
Pillow==8.1.2
-e ./plugin_example
